+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(stocklion)
 
 set(CMAKE_CXX_STANDARD 11)
 
-add_executable(stocklion main.cpp datastructures/Piece.cpp datastructures/Piece.h datastructures/Pawn.cpp datastructures/Pawn.h datastructures/Knight.cpp datastructures/Knight.h datastructures/Rook.cpp datastructures/Rook.h datastructures/Bishop.cpp datastructures/Bishop.h datastructures/King.cpp datastructures/King.h datastructures/Queen.cpp datastructures/Queen.h)
+add_executable(stocklion main.cpp datastructures/pieces/Piece.cpp datastructures/pieces/Piece.h datastructures/pieces/Pawn.cpp datastructures/pieces/Pawn.h datastructures/pieces/Knight.cpp datastructures/pieces/Knight.h datastructures/pieces/Rook.cpp datastructures/pieces/Rook.h datastructures/pieces/Bishop.cpp datastructures/pieces/Bishop.h datastructures/pieces/King.cpp datastructures/pieces/King.h datastructures/pieces/Queen.cpp datastructures/pieces/Queen.h)
