+++ datastructures/pieces/Bishop.cpp
