+++ datastructures/pieces/Bishop.h
