+++ datastructures/pieces/Queen.cpp
