+++ datastructures/pieces/Queen.h
